LayoutTests/storage/indexeddb/keypath-basics-expected.txt

Test for valid and invalid keypaths

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
globalKeyPath = 'null'
db.createObjectStore('name', {keyPath: globalKeyPath})
Deleted all object stores.
globalKeyPath = 'undefined'
db.createObjectStore('name', {keyPath: globalKeyPath})
Deleted all object stores.
globalKeyPath = ''
db.createObjectStore('name', {keyPath: globalKeyPath})
Deleted all object stores.
globalKeyPath = 'foo'
db.createObjectStore('name', {keyPath: globalKeyPath})
Deleted all object stores.
globalKeyPath = 'foo.bar.baz'
db.createObjectStore('name', {keyPath: globalKeyPath})
Deleted all object stores.
globalKeyPath = 'null'
store = db.createObjectStore('storeName')
store.createIndex('name', globalKeyPath)
Deleted all object stores.
globalKeyPath = 'undefined'
store = db.createObjectStore('storeName')
store.createIndex('name', globalKeyPath)
Deleted all object stores.
globalKeyPath = ''
store = db.createObjectStore('storeName')
store.createIndex('name', globalKeyPath)
Deleted all object stores.
globalKeyPath = 'foo'
store = db.createObjectStore('storeName')
store.createIndex('name', globalKeyPath)
Deleted all object stores.
globalKeyPath = 'foo.bar.baz'
store = db.createObjectStore('storeName')
store.createIndex('name', globalKeyPath)
Deleted all object stores.
Deleted all object stores.
globalKeyPath = '[]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo"]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo", "bar"]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["", ""]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '[1.0, 2.0]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '[["foo"]]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo", ["bar"]]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = ' '
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo '
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo. bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo .bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo..bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '+foo'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo%'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = ' '
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo '
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo. bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo .bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo..bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '+foo'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo%'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
PASS successfullyParsed is true

TEST COMPLETE
